requirements.txt
python-telegram-bot==13.15
selenium==4.10.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
emoji==2.10.0
python-dotenv==1.0.1
groq
